requirements.txt
2captcha-python==1.5.1
aioconsole==0.8.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
decorator==5.2.1
distro==1.9.0
dotenv==0.9.9
et_xmlfile==2.0.0
execnet==2.1.1
frozenlist==1.7.0
git-filter-repo==2.47.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.3
groq==0.30.0
grpcio==1.74.0
grpcio-status==1.71.2
gspread==6.2.1
gspread-dataframe==4.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.6.3
numpy==2.3.2
oauthlib==3.3.1
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
playwright==1.54.0
playwright-stealth==2.0.0
pluggy==1.6.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.2
pyparsing==3.2.3
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-base-url==2.1.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-playwright==0.7.0
pytest-rerunfailures==16.0.1
pytest-timeout==2.4.0
pytest-tornasync==0.6.0.post2
pytest-trio==0.8.0
pytest-twisted==1.14.3
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
pytz==2025.2
requests==2.32.4
requests-oauthlib==2.0.0
respx==0.22.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
text-unidecode==1.3
tornado==6.5.2
tqdm==4.67.1
trio==0.30.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
yarl==1.20.1
